Simplify API, fix global scope
Simplify API and remove chainl and chainr combinators because
they was untested anyway.
Polish naming of some API method to be more closer to original
Parsimmon ones.
Start adding operator overloading for Parser abstract for nice
syntastic sugar. Because of this, Parsihax do not need to
depend on Monax anymore.
Resolve #1 issue, that was related to Parsihax residing in global
scope.